requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
gunicorn==21.2.0
werkzeug==2.3.7
python-dotenv==1.0.0